Remove non-US layouts from keyboard package
Also simplifies back the modules for a smaller diff,
and adjusts and fixes some smaller points.
