--- packages/keyboard/src/layouts/fr-FR.ts
@@ -1,117 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-import { IKeyboardLayout, KeycodeLayout } from '../core';
-
-import { MODIFIER_KEYS } from '../special-keys';
-
-/**
- * A code-based keyboard layout for a common Norwegian keyboard.
- *
- * Note that this does not include Apple's magic Keyboards, as they map
- * the keys next to the Enter key differently (BracketRight and
- * Backslash on en-US).
- */
-export const FR_FR: IKeyboardLayout = new KeycodeLayout(
-  'fr-FR',
-  {},
-  MODIFIER_KEYS,
-  {
-    AltLeft: 'Alt',
-    AltRight: 'AltGraph',
-    ArrowDown: 'ArrowDown',
-    ArrowLeft: 'ArrowLeft',
-    ArrowRight: 'ArrowRight',
-    ArrowUp: 'ArrowUp',
-    Backquote: '²',
-    Backslash: '*',
-    Backspace: 'Backspace',
-    BracketRight: '$',
-    Comma: ';',
-    ControlLeft: 'Control',
-    ControlRight: 'Control',
-    Delete: 'Delete',
-    Digit0: 'À',
-    Digit1: '&',
-    Digit2: 'É',
-    Digit3: '"',
-    Digit4: "'",
-    Digit5: '(',
-    Digit6: '-',
-    Digit7: 'È',
-    Digit8: '_',
-    Digit9: 'Ç',
-    End: 'End',
-    Enter: 'Enter',
-    Equal: '=',
-    Escape: 'Escape',
-    F1: 'F1',
-    F10: 'F10',
-    F11: 'F11',
-    F12: 'F12',
-    F2: 'F2',
-    F3: 'F3',
-    F4: 'F4',
-    F5: 'F5',
-    F6: 'F6',
-    F7: 'F7',
-    F8: 'F8',
-    F9: 'F9',
-    Home: 'Home',
-    Insert: 'Insert',
-    IntlBackslash: '<',
-    KeyA: 'Q',
-    KeyB: 'B',
-    KeyC: 'C',
-    KeyD: 'D',
-    KeyE: 'E',
-    KeyF: 'F',
-    KeyG: 'G',
-    KeyH: 'H',
-    KeyI: 'I',
-    KeyJ: 'J',
-    KeyK: 'K',
-    KeyL: 'L',
-    KeyM: ',',
-    KeyN: 'N',
-    KeyO: 'O',
-    KeyP: 'P',
-    KeyQ: 'A',
-    KeyR: 'R',
-    KeyS: 'S',
-    KeyT: 'T',
-    KeyU: 'U',
-    KeyV: 'V',
-    KeyW: 'Z',
-    KeyX: 'X',
-    KeyY: 'Y',
-    KeyZ: 'W',
-    Minus: ')',
-    Numpad0: '0',
-    Numpad1: '1',
-    Numpad2: '2',
-    Numpad3: '3',
-    Numpad4: '4',
-    Numpad5: '5',
-    Numpad6: '6',
-    Numpad7: '7',
-    Numpad8: '8',
-    Numpad9: '9',
-    NumpadAdd: '+',
-    NumpadDecimal: '.',
-    NumpadDivide: '/',
-    NumpadEnter: 'Enter',
-    NumpadMultiply: '*',
-    NumpadSubtract: '-',
-    PageDown: 'PageDown',
-    PageUp: 'PageUp',
-    Period: ':',
-    Quote: 'Ù',
-    ScrollLock: 'ScrollLock',
-    Semicolon: 'M',
-    ShiftLeft: 'Shift',
-    ShiftRight: 'Shift',
-    Slash: '!',
-    Tab: 'Tab'
-  }
-);
