--- packages/keyboard/src/layouts/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-export { EN_US } from './en-US';
-export { FR_FR } from './fr-FR';
-export { NB_NO } from './nb-NO';
