--- packages/keyboard/src/layouts/nb-NO.ts
@@ -1,123 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-import { IKeyboardLayout, KeycodeLayout } from '../core';
-
-import { MODIFIER_KEYS } from '../special-keys';
-
-/**
- * A code-based keyboard layout for a common Norwegian keyboard.
- *
- * Note that this does not include Apple's magic Keyboards, as they map
- * the keys next to the Enter key differently (BracketRight and
- * Backslash on en-US).
- */
-export const NB_NO: IKeyboardLayout = new KeycodeLayout(
-  'nb-NO',
-  {},
-  MODIFIER_KEYS,
-  {
-    AltLeft: 'Alt',
-    AltRight: 'AltGraph',
-    Backquote: '|',
-    Backslash: "'",
-    Backspace: 'Backspace',
-    BracketLeft: 'Å',
-    CapsLock: 'CapsLock',
-    Comma: ',',
-    ContextMenu: 'ContextMenu',
-    ControlLeft: 'Control',
-    ControlRight: 'Control',
-    Delete: 'Delete',
-    Digit0: '0',
-    Digit1: '1',
-    Digit2: '2',
-    Digit3: '3',
-    Digit4: '4',
-    Digit5: '5',
-    Digit6: '6',
-    Digit7: '7',
-    Digit8: '8',
-    Digit9: '9',
-    End: 'End',
-    Enter: 'Enter',
-    Equal: '\\',
-    Escape: 'Escape',
-    F1: 'F1',
-    F10: 'F10',
-    F11: 'F11',
-    F12: 'F12',
-    F2: 'F2',
-    F3: 'F3',
-    F4: 'F4',
-    F5: 'F5',
-    F6: 'F6',
-    F7: 'F7',
-    F8: 'F8',
-    F9: 'F9',
-    Home: 'Home',
-    Insert: 'Insert',
-    IntlBackslash: '<',
-    KeyA: 'A',
-    KeyB: 'B',
-    KeyC: 'C',
-    KeyD: 'D',
-    KeyE: 'E',
-    KeyF: 'F',
-    KeyG: 'G',
-    KeyH: 'H',
-    KeyI: 'I',
-    KeyJ: 'J',
-    KeyK: 'K',
-    KeyL: 'L',
-    KeyM: 'M',
-    KeyN: 'N',
-    KeyO: 'O',
-    KeyP: 'P',
-    KeyQ: 'Q',
-    KeyR: 'R',
-    KeyS: 'S',
-    KeyT: 'T',
-    KeyU: 'U',
-    KeyV: 'V',
-    KeyW: 'W',
-    KeyX: 'X',
-    KeyY: 'Y',
-    KeyZ: 'Z',
-    MetaLeft: 'Meta', // chrome
-    MetaRight: 'Meta', // chrome
-    Minus: '+',
-    NumLock: 'NumLock',
-    Numpad0: 'Insert',
-    Numpad1: 'End',
-    Numpad2: 'ArrowDown',
-    Numpad3: 'PageDown',
-    Numpad4: 'ArrowLeft',
-    Numpad5: 'Clear',
-    Numpad6: 'ArrowRight',
-    Numpad7: 'Home',
-    Numpad8: 'ArrowUp',
-    Numpad9: 'PageUp',
-    NumpadAdd: '+',
-    NumpadDecimal: 'Delete',
-    NumpadDivide: '/',
-    NumpadEnter: 'Enter',
-    NumpadMultiply: '*',
-    NumpadSubtract: '-',
-    OSLeft: 'OS', // firefox
-    OSRight: 'OS', // firefox
-    PageDown: 'PageDown',
-    PageUp: 'PageUp',
-    Pause: 'Pause',
-    Period: '.',
-    PrintScreen: 'PrintScreen',
-    Quote: 'Æ',
-    ScrollLock: 'ScrollLock',
-    Semicolon: 'Ø',
-    ShiftLeft: 'Shift',
-    ShiftRight: 'Shift',
-    Slash: '-',
-    Space: ' ',
-    Tab: 'Tab'
-  }
-);
