+++ packages/keyboard/tests/src/index.spec.ts
@@ -12,13 +12,238 @@ import { expect } from 'chai';
 import {
   EN_US,
   getKeyboardLayout,
-  KeyboardLayouts,
   KeycodeLayout,
   setKeyboardLayout
 } from '@lumino/keyboard';
 
-const FR_FR = KeyboardLayouts.find(layout => layout.name === 'fr-FR');
-const NB_NO = KeyboardLayouts.find(layout => layout.name === 'nb-NO');
+const MODIFIER_KEYS = [
+  'Alt',
+  'AltGraph',
+  'CapsLock',
+  'Control',
+  'Fn',
+  'FnLock',
+  'Meta',
+  'NumLock',
+  'ScrollLock',
+  'Shift',
+  'Symbol',
+  'SymbolLock'
+];
+
+/**
+ * A common Norwegian keyboard layout.
+ *
+ * Note that this does not include Apple's magic Keyboards, as they map
+ * the keys next to the Enter key differently (BracketRight and
+ * Backslash on en-US).
+ */
+export const NB_NO = new KeycodeLayout('nb-NO', {}, MODIFIER_KEYS, {
+  AltLeft: 'Alt',
+  AltRight: 'AltGraph',
+  Backquote: '|',
+  Backslash: "'",
+  Backspace: 'Backspace',
+  BracketLeft: 'Å',
+  CapsLock: 'CapsLock',
+  Comma: ',',
+  ContextMenu: 'ContextMenu',
+  ControlLeft: 'Control',
+  ControlRight: 'Control',
+  Delete: 'Delete',
+  Digit0: '0',
+  Digit1: '1',
+  Digit2: '2',
+  Digit3: '3',
+  Digit4: '4',
+  Digit5: '5',
+  Digit6: '6',
+  Digit7: '7',
+  Digit8: '8',
+  Digit9: '9',
+  End: 'End',
+  Enter: 'Enter',
+  Equal: '\\',
+  Escape: 'Escape',
+  F1: 'F1',
+  F10: 'F10',
+  F11: 'F11',
+  F12: 'F12',
+  F2: 'F2',
+  F3: 'F3',
+  F4: 'F4',
+  F5: 'F5',
+  F6: 'F6',
+  F7: 'F7',
+  F8: 'F8',
+  F9: 'F9',
+  Home: 'Home',
+  Insert: 'Insert',
+  IntlBackslash: '<',
+  KeyA: 'A',
+  KeyB: 'B',
+  KeyC: 'C',
+  KeyD: 'D',
+  KeyE: 'E',
+  KeyF: 'F',
+  KeyG: 'G',
+  KeyH: 'H',
+  KeyI: 'I',
+  KeyJ: 'J',
+  KeyK: 'K',
+  KeyL: 'L',
+  KeyM: 'M',
+  KeyN: 'N',
+  KeyO: 'O',
+  KeyP: 'P',
+  KeyQ: 'Q',
+  KeyR: 'R',
+  KeyS: 'S',
+  KeyT: 'T',
+  KeyU: 'U',
+  KeyV: 'V',
+  KeyW: 'W',
+  KeyX: 'X',
+  KeyY: 'Y',
+  KeyZ: 'Z',
+  MetaLeft: 'Meta', // chrome
+  MetaRight: 'Meta', // chrome
+  Minus: '+',
+  NumLock: 'NumLock',
+  Numpad0: 'Insert',
+  Numpad1: 'End',
+  Numpad2: 'ArrowDown',
+  Numpad3: 'PageDown',
+  Numpad4: 'ArrowLeft',
+  Numpad5: 'Clear',
+  Numpad6: 'ArrowRight',
+  Numpad7: 'Home',
+  Numpad8: 'ArrowUp',
+  Numpad9: 'PageUp',
+  NumpadAdd: '+',
+  NumpadDecimal: 'Delete',
+  NumpadDivide: '/',
+  NumpadEnter: 'Enter',
+  NumpadMultiply: '*',
+  NumpadSubtract: '-',
+  OSLeft: 'OS', // firefox
+  OSRight: 'OS', // firefox
+  PageDown: 'PageDown',
+  PageUp: 'PageUp',
+  Pause: 'Pause',
+  Period: '.',
+  PrintScreen: 'PrintScreen',
+  Quote: 'Æ',
+  ScrollLock: 'ScrollLock',
+  Semicolon: 'Ø',
+  ShiftLeft: 'Shift',
+  ShiftRight: 'Shift',
+  Slash: '-',
+  Space: ' ',
+  Tab: 'Tab'
+});
+
+/**
+ * A common French keyboard layout
+ */
+export const FR_FR = new KeycodeLayout('fr-FR', {}, MODIFIER_KEYS, {
+  AltLeft: 'Alt',
+  AltRight: 'AltGraph',
+  ArrowDown: 'ArrowDown',
+  ArrowLeft: 'ArrowLeft',
+  ArrowRight: 'ArrowRight',
+  ArrowUp: 'ArrowUp',
+  Backquote: '²',
+  Backslash: '*',
+  Backspace: 'Backspace',
+  BracketRight: '$',
+  Comma: ';',
+  ControlLeft: 'Control',
+  ControlRight: 'Control',
+  Delete: 'Delete',
+  Digit0: 'À',
+  Digit1: '&',
+  Digit2: 'É',
+  Digit3: '"',
+  Digit4: "'",
+  Digit5: '(',
+  Digit6: '-',
+  Digit7: 'È',
+  Digit8: '_',
+  Digit9: 'Ç',
+  End: 'End',
+  Enter: 'Enter',
+  Equal: '=',
+  Escape: 'Escape',
+  F1: 'F1',
+  F10: 'F10',
+  F11: 'F11',
+  F12: 'F12',
+  F2: 'F2',
+  F3: 'F3',
+  F4: 'F4',
+  F5: 'F5',
+  F6: 'F6',
+  F7: 'F7',
+  F8: 'F8',
+  F9: 'F9',
+  Home: 'Home',
+  Insert: 'Insert',
+  IntlBackslash: '<',
+  KeyA: 'Q',
+  KeyB: 'B',
+  KeyC: 'C',
+  KeyD: 'D',
+  KeyE: 'E',
+  KeyF: 'F',
+  KeyG: 'G',
+  KeyH: 'H',
+  KeyI: 'I',
+  KeyJ: 'J',
+  KeyK: 'K',
+  KeyL: 'L',
+  KeyM: ',',
+  KeyN: 'N',
+  KeyO: 'O',
+  KeyP: 'P',
+  KeyQ: 'A',
+  KeyR: 'R',
+  KeyS: 'S',
+  KeyT: 'T',
+  KeyU: 'U',
+  KeyV: 'V',
+  KeyW: 'Z',
+  KeyX: 'X',
+  KeyY: 'Y',
+  KeyZ: 'W',
+  Minus: ')',
+  Numpad0: '0',
+  Numpad1: '1',
+  Numpad2: '2',
+  Numpad3: '3',
+  Numpad4: '4',
+  Numpad5: '5',
+  Numpad6: '6',
+  Numpad7: '7',
+  Numpad8: '8',
+  Numpad9: '9',
+  NumpadAdd: '+',
+  NumpadDecimal: '.',
+  NumpadDivide: '/',
+  NumpadEnter: 'Enter',
+  NumpadMultiply: '*',
+  NumpadSubtract: '-',
+  PageDown: 'PageDown',
+  PageUp: 'PageUp',
+  Period: ':',
+  Quote: 'Ù',
+  ScrollLock: 'ScrollLock',
+  Semicolon: 'M',
+  ShiftLeft: 'Shift',
+  ShiftRight: 'Shift',
+  Slash: '!',
+  Tab: 'Tab'
+});
 
 describe('@lumino/keyboard', () => {
   describe('getKeyboardLayout()', () => {
